Move e2e tests to independent folder (#297)
* wip

* provider

* rename test suites

* two more

* consumer

* final cleans and readme

* Update README.md

* Update README.md

* mas formatting in readme

* Update README.md

* Update README.md

* pr changes

* Update README.md


* remove debug code
