+++ .github/workflows/automated-tests.yml
@@ -11,10 +11,7 @@ jobs:
         with:
           go-version: "1.18.0" # The Go version to download (if necessary) and use.
 
-      - name: Figure out where we are
-        run: ls
-
-      - name: Unit tests
+      - name: Unit and e2e tests
         run: go test ./...
 
       - name: Integration tests
