+++ e2e-tests/stop_consumer_test.go
@@ -1,4 +1,4 @@
-package provider_test
+package e2e_test
 
 import (
 	"time"
