--- x/ccv/provider/provider_test.go
@@ -1,478 +1,0 @@
-package provider_test
-
-import (
-	"bytes"
-	"testing"
-	"time"
-
-	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	evidencetypes "github.com/cosmos/cosmos-sdk/x/evidence/types"
-	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	ccv "github.com/cosmos/interchain-security/x/ccv/types"
-	"github.com/cosmos/interchain-security/x/ccv/utils"
-
-	transfertypes "github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
-	clienttypes "github.com/cosmos/ibc-go/v3/modules/core/02-client/types"
-	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
-	ibctesting "github.com/cosmos/ibc-go/v3/testing"
-
-	appConsumer "github.com/cosmos/interchain-security/app/consumer"
-	appProvider "github.com/cosmos/interchain-security/app/provider"
-	"github.com/cosmos/interchain-security/testutil/simapp"
-	consumertypes "github.com/cosmos/interchain-security/x/ccv/consumer/types"
-	providertypes "github.com/cosmos/interchain-security/x/ccv/provider/types"
-	"github.com/cosmos/interchain-security/x/ccv/types"
-
-	abci "github.com/tendermint/tendermint/abci/types"
-	tmtypes "github.com/tendermint/tendermint/types"
-
-	"github.com/stretchr/testify/suite"
-)
-
-type ProviderTestSuite struct {
-	suite.Suite
-
-	coordinator *ibctesting.Coordinator
-
-	// testing chains
-	providerChain *ibctesting.TestChain
-	consumerChain *ibctesting.TestChain
-
-	path         *ibctesting.Path
-	transferPath *ibctesting.Path
-}
-
-func (suite *ProviderTestSuite) SetupTest() {
-	suite.coordinator, suite.providerChain, suite.consumerChain = simapp.NewProviderConsumerCoordinator(suite.T())
-
-	// valsets must match
-	providerValUpdates := tmtypes.TM2PB.ValidatorUpdates(suite.providerChain.Vals)
-	consumerValUpdates := tmtypes.TM2PB.ValidatorUpdates(suite.consumerChain.Vals)
-	suite.Require().True(len(providerValUpdates) == len(consumerValUpdates), "initial valset not matching")
-	for i := 0; i < len(providerValUpdates); i++ {
-		addr1 := utils.GetChangePubKeyAddress(providerValUpdates[i])
-		addr2 := utils.GetChangePubKeyAddress(consumerValUpdates[i])
-		suite.Require().True(bytes.Equal(addr1, addr2), "validator mismatch")
-	}
-
-	// move both chains to the next block
-	suite.providerChain.NextBlock()
-	suite.consumerChain.NextBlock()
-
-	// create consumer client on provider chain and set as consumer client for consumer chainID in provider keeper.
-	err := suite.providerChain.App.(*appProvider.App).ProviderKeeper.CreateConsumerClient(
-		suite.providerCtx(),
-		suite.consumerChain.ChainID,
-		suite.consumerChain.LastHeader.GetHeight().(clienttypes.Height),
-		false,
-	)
-	suite.Require().NoError(err)
-	// move provider to next block to commit the state
-	suite.providerChain.NextBlock()
-
-	// initialize the consumer chain with the genesis state stored on the provider
-	consumerGenesis, found := suite.providerChain.App.(*appProvider.App).ProviderKeeper.GetConsumerGenesis(
-		suite.providerCtx(),
-		suite.consumerChain.ChainID,
-	)
-	suite.Require().True(found, "consumer genesis not found")
-	suite.consumerChain.App.(*appConsumer.App).ConsumerKeeper.InitGenesis(suite.consumerChain.GetContext(), &consumerGenesis)
-
-	// create path for the CCV channel
-	suite.path = ibctesting.NewPath(suite.consumerChain, suite.providerChain)
-
-	// update CCV path with correct info
-	// - set provider endpoint's clientID
-	consumerClient, found := suite.providerChain.App.(*appProvider.App).ProviderKeeper.GetConsumerClientId(
-		suite.providerCtx(),
-		suite.consumerChain.ChainID,
-	)
-	suite.Require().True(found, "consumer client not found")
-	suite.path.EndpointB.ClientID = consumerClient
-	// - set consumer endpoint's clientID
-	providerClient, found := suite.consumerChain.App.(*appConsumer.App).ConsumerKeeper.GetProviderClientID(suite.consumerChain.GetContext())
-	suite.Require().True(found, "provider client not found")
-	suite.path.EndpointA.ClientID = providerClient
-	// - client config
-	providerUnbondingPeriod := suite.providerChain.App.(*appProvider.App).GetStakingKeeper().UnbondingTime(suite.providerCtx())
-	suite.path.EndpointB.ClientConfig.(*ibctesting.TendermintConfig).UnbondingPeriod = providerUnbondingPeriod
-	suite.path.EndpointB.ClientConfig.(*ibctesting.TendermintConfig).TrustingPeriod = providerUnbondingPeriod / utils.TrustingPeriodFraction
-	consumerUnbondingPeriod := utils.ComputeConsumerUnbondingPeriod(providerUnbondingPeriod)
-	suite.path.EndpointA.ClientConfig.(*ibctesting.TendermintConfig).UnbondingPeriod = consumerUnbondingPeriod
-	suite.path.EndpointA.ClientConfig.(*ibctesting.TendermintConfig).TrustingPeriod = consumerUnbondingPeriod / utils.TrustingPeriodFraction
-	// - channel config
-	suite.path.EndpointA.ChannelConfig.PortID = consumertypes.PortID
-	suite.path.EndpointB.ChannelConfig.PortID = providertypes.PortID
-	suite.path.EndpointA.ChannelConfig.Version = types.Version
-	suite.path.EndpointB.ChannelConfig.Version = types.Version
-	suite.path.EndpointA.ChannelConfig.Order = channeltypes.ORDERED
-	suite.path.EndpointB.ChannelConfig.Order = channeltypes.ORDERED
-
-	// set chains sender account number
-	// TODO: to be fixed in #151
-	err = suite.path.EndpointB.Chain.SenderAccount.SetAccountNumber(6)
-	suite.Require().NoError(err)
-	err = suite.path.EndpointA.Chain.SenderAccount.SetAccountNumber(1)
-	suite.Require().NoError(err)
-
-	// create path for the transfer channel
-	suite.transferPath = ibctesting.NewPath(suite.consumerChain, suite.providerChain)
-	suite.transferPath.EndpointA.ChannelConfig.PortID = transfertypes.PortID
-	suite.transferPath.EndpointB.ChannelConfig.PortID = transfertypes.PortID
-	suite.transferPath.EndpointA.ChannelConfig.Version = transfertypes.Version
-	suite.transferPath.EndpointB.ChannelConfig.Version = transfertypes.Version
-}
-
-func (suite *ProviderTestSuite) SetupCCVChannel() {
-	suite.StartSetupCCVChannel()
-	suite.CompleteSetupCCVChannel()
-	suite.SetupTransferChannel()
-}
-
-func (suite *ProviderTestSuite) StartSetupCCVChannel() {
-	suite.coordinator.CreateConnections(suite.path)
-
-	err := suite.path.EndpointA.ChanOpenInit()
-	suite.Require().NoError(err)
-
-	err = suite.path.EndpointB.ChanOpenTry()
-	suite.Require().NoError(err)
-}
-
-func (suite *ProviderTestSuite) CompleteSetupCCVChannel() {
-	err := suite.path.EndpointA.ChanOpenAck()
-	suite.Require().NoError(err)
-
-	err = suite.path.EndpointB.ChanOpenConfirm()
-	suite.Require().NoError(err)
-
-	// ensure counterparty is up to date
-	err = suite.path.EndpointA.UpdateClient()
-	suite.Require().NoError(err)
-}
-
-func (suite *ProviderTestSuite) SetupTransferChannel() {
-	// transfer path will use the same connection as ccv path
-
-	suite.transferPath.EndpointA.ClientID = suite.path.EndpointA.ClientID
-	suite.transferPath.EndpointA.ConnectionID = suite.path.EndpointA.ConnectionID
-	suite.transferPath.EndpointB.ClientID = suite.path.EndpointB.ClientID
-	suite.transferPath.EndpointB.ConnectionID = suite.path.EndpointB.ConnectionID
-
-	// CCV channel handshake will automatically initiate transfer channel handshake on ACK
-	// so transfer channel will be on stage INIT when CompleteSetupCCVChannel returns.
-	suite.transferPath.EndpointA.ChannelID = suite.consumerChain.App.(*appConsumer.App).
-		ConsumerKeeper.GetDistributionTransmissionChannel(suite.consumerChain.GetContext())
-
-	// Complete TRY, ACK, CONFIRM for transfer path
-	err := suite.transferPath.EndpointB.ChanOpenTry()
-	suite.Require().NoError(err)
-
-	err = suite.transferPath.EndpointA.ChanOpenAck()
-	suite.Require().NoError(err)
-
-	err = suite.transferPath.EndpointB.ChanOpenConfirm()
-	suite.Require().NoError(err)
-
-	// ensure counterparty is up to date
-	err = suite.transferPath.EndpointA.UpdateClient()
-	suite.Require().NoError(err)
-}
-
-func TestProviderTestSuite(t *testing.T) {
-	suite.Run(t, new(ProviderTestSuite))
-}
-
-func (s *ProviderTestSuite) TestPacketRoundtrip() {
-	s.SetupCCVChannel()
-
-	// Bond some tokens on provider to change validator powers
-	bondAmt := sdk.NewInt(1000000)
-	delAddr := s.providerChain.SenderAccount.GetAddress()
-	delegate(s, delAddr, bondAmt)
-
-	// Send CCV packet to consumer
-	s.providerChain.NextBlock()
-
-	// Relay 1 VSC packet from provider to consumer
-	relayAllCommittedPackets(s, s.providerChain, s.path, providertypes.PortID, s.path.EndpointB.ChannelID, 1)
-
-	// Increment time so that the unbonding period ends on the provider
-	incrementTimeByUnbondingPeriod(s, Provider)
-
-	// Relay 1 VSCMatured packet from consumer to provider
-	relayAllCommittedPackets(s, s.consumerChain, s.path, consumertypes.PortID, s.path.EndpointA.ChannelID, 1)
-}
-
-func (s *ProviderTestSuite) providerCtx() sdk.Context {
-	return s.providerChain.GetContext()
-}
-
-func (s *ProviderTestSuite) consumerCtx() sdk.Context {
-	return s.consumerChain.GetContext()
-}
-
-func (s *ProviderTestSuite) providerBondDenom() string {
-	return s.providerChain.App.(*appProvider.App).StakingKeeper.BondDenom(s.providerCtx())
-}
-
-// TestSendDowntimePacket tests consumer initiated slashing
-func (s *ProviderTestSuite) TestSendSlashPacketDowntime() {
-	s.SetupCCVChannel()
-	validatorsPerChain := len(s.consumerChain.Vals.Validators)
-
-	providerStakingKeeper := s.providerChain.App.(*appProvider.App).StakingKeeper
-	providerSlashingKeeper := s.providerChain.App.(*appProvider.App).SlashingKeeper
-
-	consumerKeeper := s.consumerChain.App.(*appConsumer.App).ConsumerKeeper
-
-	// get a cross-chain validator address, pubkey and balance
-	tmVals := s.consumerChain.Vals.Validators
-	tmVal := tmVals[0]
-
-	val, err := tmVal.ToProto()
-	s.Require().NoError(err)
-	pubkey, err := cryptocodec.FromTmProtoPublicKey(val.GetPubKey())
-	s.Require().Nil(err)
-	consAddr := sdk.GetConsAddress(pubkey)
-	valData, found := providerStakingKeeper.GetValidatorByConsAddr(s.providerCtx(), consAddr)
-	s.Require().True(found)
-	valOldBalance := valData.Tokens
-
-	// create the validator's signing info record to allow jailing
-	valInfo := slashingtypes.NewValidatorSigningInfo(consAddr, s.providerCtx().BlockHeight(),
-		s.providerCtx().BlockHeight()-1, time.Time{}.UTC(), false, int64(0))
-	providerSlashingKeeper.SetValidatorSigningInfo(s.providerCtx(), consAddr, valInfo)
-
-	// get valseUpdateId for current block height
-	valsetUpdateId := consumerKeeper.GetHeightValsetUpdateID(s.consumerCtx(), uint64(s.consumerCtx().BlockHeight()))
-
-	// construct the downtime packet with the validator address and power along
-	// with the slashing and jailing parameters
-	validator := abci.Validator{
-		Address: tmVal.Address,
-		Power:   tmVal.VotingPower,
-	}
-
-	oldBlockTime := s.consumerCtx().BlockTime()
-	slashFraction := int64(100)
-	packetData := types.NewSlashPacketData(validator, valsetUpdateId, stakingtypes.Downtime)
-	timeout := uint64(types.GetTimeoutTimestamp(oldBlockTime).UnixNano())
-	packet := channeltypes.NewPacket(packetData.GetBytes(), 1, consumertypes.PortID, s.path.EndpointA.ChannelID,
-		providertypes.PortID, s.path.EndpointB.ChannelID, clienttypes.Height{}, timeout)
-
-	// Send the downtime packet through CCV
-	err = s.path.EndpointA.SendPacket(packet)
-	s.Require().NoError(err)
-
-	// Set outstanding slashing flag
-	s.consumerChain.App.(*appConsumer.App).ConsumerKeeper.SetOutstandingDowntime(s.consumerCtx(), consAddr)
-
-	// save next VSC packet info
-	oldBlockTime = s.providerCtx().BlockTime()
-	timeout = uint64(types.GetTimeoutTimestamp(oldBlockTime).UnixNano())
-	valsetUpdateID := s.providerChain.App.(*appProvider.App).ProviderKeeper.GetValidatorSetUpdateId(s.providerCtx())
-
-	// receive the downtime packet on the provider chain;
-	// RecvPacket() calls the provider endblocker thus sends a VSC packet to the consumer
-	err = s.path.EndpointB.RecvPacket(packet)
-	s.Require().NoError(err)
-
-	// check that the validator was removed from the provider validator set
-	s.Require().Len(s.providerChain.Vals.Validators, validatorsPerChain-1)
-	// check that the VSC ID is updated on the consumer chain
-
-	// update consumer client on the VSC packet sent from provider
-	err = s.path.EndpointA.UpdateClient()
-	s.Require().NoError(err)
-
-	// reconstruct VSC packet
-	valUpdates := []abci.ValidatorUpdate{
-		{
-			PubKey: val.GetPubKey(),
-			Power:  int64(0),
-		},
-	}
-	packetData2 := ccv.NewValidatorSetChangePacketData(valUpdates, valsetUpdateID, []string{consAddr.String()})
-	packet2 := channeltypes.NewPacket(packetData2.GetBytes(), 1, providertypes.PortID, s.path.EndpointB.ChannelID,
-		consumertypes.PortID, s.path.EndpointA.ChannelID, clienttypes.Height{}, timeout)
-
-	// receive VSC packet about jailing on the consumer chain
-	err = s.path.EndpointA.RecvPacket(packet2)
-	s.Require().NoError(err)
-
-	// check that the consumer update its VSC ID for the subsequent block
-	s.Require().Equal(consumerKeeper.GetHeightValsetUpdateID(s.consumerCtx(), uint64(s.consumerCtx().BlockHeight())+1), valsetUpdateID)
-
-	// check that the validator was removed from the consumer validator set
-	s.Require().Len(s.consumerChain.Vals.Validators, validatorsPerChain-1)
-
-	err = s.path.EndpointB.UpdateClient()
-	s.Require().NoError(err)
-
-	// check that the validator is successfully jailed on provider
-
-	validatorJailed, ok := s.providerChain.App.(*appProvider.App).StakingKeeper.GetValidatorByConsAddr(s.providerCtx(), consAddr)
-	s.Require().True(ok)
-	s.Require().True(validatorJailed.Jailed)
-	s.Require().Equal(validatorJailed.Status, stakingtypes.Unbonding)
-
-	// check that the validator's token was slashed
-	slashedAmout := sdk.NewDec(1).QuoInt64(slashFraction).Mul(valOldBalance.ToDec())
-	resultingTokens := valOldBalance.Sub(slashedAmout.TruncateInt())
-	s.Require().Equal(resultingTokens, validatorJailed.GetTokens())
-
-	// check that the validator's unjailing time is updated
-	valSignInfo, found := providerSlashingKeeper.GetValidatorSigningInfo(s.providerCtx(), consAddr)
-	s.Require().True(found)
-	s.Require().True(valSignInfo.JailedUntil.After(s.providerCtx().BlockHeader().Time))
-
-	// check that the outstanding slashing flag is reset on the consumer
-	pFlag := s.consumerChain.App.(*appConsumer.App).ConsumerKeeper.OutstandingDowntime(s.consumerCtx(), consAddr)
-	s.Require().False(pFlag)
-
-	// check that slashing packet gets acknowledged
-	ack := channeltypes.NewResultAcknowledgement([]byte{byte(1)})
-	err = s.path.EndpointA.AcknowledgePacket(packet, ack.Acknowledgement())
-	s.Require().NoError(err)
-}
-
-func (s *ProviderTestSuite) TestSendSlashPacketDoubleSign() {
-	s.SetupCCVChannel()
-	validatorsPerChain := len(s.consumerChain.Vals.Validators)
-
-	providerStakingKeeper := s.providerChain.App.(*appProvider.App).StakingKeeper
-	providerSlashingKeeper := s.providerChain.App.(*appProvider.App).SlashingKeeper
-	consumerKeeper := s.consumerChain.App.(*appConsumer.App).ConsumerKeeper
-
-	// get a cross-chain validator address, pubkey and balance
-	tmVals := s.consumerChain.Vals.Validators
-	tmVal := tmVals[0]
-
-	val, err := tmVal.ToProto()
-	s.Require().NoError(err)
-	pubkey, err := cryptocodec.FromTmProtoPublicKey(val.GetPubKey())
-	s.Require().Nil(err)
-	consAddr := sdk.GetConsAddress(pubkey)
-	valData, found := providerStakingKeeper.GetValidatorByConsAddr(s.providerCtx(), consAddr)
-	s.Require().True(found)
-	valOldBalance := valData.Tokens
-
-	// create the validator's signing info record to allow jailing
-	valInfo := slashingtypes.NewValidatorSigningInfo(consAddr, s.providerCtx().BlockHeight(),
-		s.providerCtx().BlockHeight()-1, time.Time{}.UTC(), false, int64(0))
-	providerSlashingKeeper.SetValidatorSigningInfo(s.providerCtx(), consAddr, valInfo)
-
-	// get valseUpdateId for current block height
-	valsetUpdateId := consumerKeeper.GetHeightValsetUpdateID(s.consumerCtx(), uint64(s.consumerCtx().BlockHeight()))
-
-	// construct the downtime packet with the validator address and power along
-	// with the slashing and jailing parameters
-	validator := abci.Validator{
-		Address: tmVal.Address,
-		Power:   tmVal.VotingPower,
-	}
-
-	oldBlockTime := s.consumerCtx().BlockTime()
-	packetData := types.NewSlashPacketData(validator, valsetUpdateId, stakingtypes.DoubleSign)
-
-	timeout := uint64(types.GetTimeoutTimestamp(oldBlockTime).UnixNano())
-	packet := channeltypes.NewPacket(packetData.GetBytes(), 1, consumertypes.PortID, s.path.EndpointA.ChannelID,
-		providertypes.PortID, s.path.EndpointB.ChannelID, clienttypes.Height{}, timeout)
-
-	// Send the downtime packet through CCV
-	err = s.path.EndpointA.SendPacket(packet)
-	s.Require().NoError(err)
-
-	// save next VSC packet info
-	oldBlockTime = s.providerCtx().BlockTime()
-	timeout = uint64(types.GetTimeoutTimestamp(oldBlockTime).UnixNano())
-	valsetUpdateID := s.providerChain.App.(*appProvider.App).ProviderKeeper.GetValidatorSetUpdateId(s.providerCtx())
-
-	// receive the downtime packet on the provider chain;
-	// RecvPacket() calls the provider endblocker and thus sends a VSC packet to the consumer
-	err = s.path.EndpointB.RecvPacket(packet)
-	s.Require().NoError(err)
-
-	// check that the validator was removed from the provider validator set
-	s.Require().Len(s.providerChain.Vals.Validators, validatorsPerChain-1)
-	// check that the VSC ID is updated on the consumer chain
-
-	// update consumer client on the VSC packet sent from provider
-	err = s.path.EndpointA.UpdateClient()
-	s.Require().NoError(err)
-
-	// reconstruct VSC packet
-	valUpdates := []abci.ValidatorUpdate{
-		{
-			PubKey: val.GetPubKey(),
-			Power:  int64(0),
-		},
-	}
-	packetData2 := ccv.NewValidatorSetChangePacketData(valUpdates, valsetUpdateID, []string{})
-	packet2 := channeltypes.NewPacket(packetData2.GetBytes(), 1, providertypes.PortID, s.path.EndpointB.ChannelID,
-		consumertypes.PortID, s.path.EndpointA.ChannelID, clienttypes.Height{}, timeout)
-
-	// receive VSC packet about jailing on the consumer chain
-	err = s.path.EndpointA.RecvPacket(packet2)
-	s.Require().NoError(err)
-
-	// check that the consumer update its VSC ID for the subsequent block
-	s.Require().Equal(consumerKeeper.GetHeightValsetUpdateID(s.consumerCtx(), uint64(s.consumerCtx().BlockHeight())+1), valsetUpdateID)
-
-	// check that the validator was removed from the consumer validator set
-	s.Require().Len(s.consumerChain.Vals.Validators, validatorsPerChain-1)
-
-	err = s.path.EndpointB.UpdateClient()
-	s.Require().NoError(err)
-
-	// check that the validator is successfully jailed on provider
-	validatorJailed, ok := s.providerChain.App.(*appProvider.App).StakingKeeper.GetValidatorByConsAddr(s.providerCtx(), consAddr)
-	s.Require().True(ok)
-	s.Require().True(validatorJailed.Jailed)
-	s.Require().Equal(validatorJailed.Status, stakingtypes.Unbonding)
-
-	// check that the validator's token was slashed
-	slashedAmout := providerSlashingKeeper.SlashFractionDoubleSign(s.providerCtx()).Mul(valOldBalance.ToDec())
-	resultingTokens := valOldBalance.Sub(slashedAmout.TruncateInt())
-	s.Require().Equal(resultingTokens, validatorJailed.GetTokens())
-
-	// check that the validator's unjailing time is updated
-	valSignInfo, found := providerSlashingKeeper.GetValidatorSigningInfo(s.providerCtx(), consAddr)
-	s.Require().True(found)
-	s.Require().True(valSignInfo.JailedUntil.After(s.providerCtx().BlockHeader().Time))
-
-	// check that validator was tombstoned
-	s.Require().True(valSignInfo.Tombstoned)
-	s.Require().True(valSignInfo.JailedUntil.Equal(evidencetypes.DoubleSignJailEndTime))
-}
-
-func (s *ProviderTestSuite) getVal(index int) (validator stakingtypes.Validator, valAddr sdk.ValAddress) {
-	// Choose a validator, and get its address and data structure into the correct types
-	tmValidator := s.providerChain.Vals.Validators[index]
-	valAddr, err := sdk.ValAddressFromHex(tmValidator.Address.String())
-	s.Require().NoError(err)
-	validator, found := s.providerChain.App.(*appProvider.App).StakingKeeper.GetValidator(s.providerCtx(), valAddr)
-	s.Require().True(found)
-
-	return validator, valAddr
-}
-
-func (s *ProviderTestSuite) TestSlashPacketAcknowldgement() {
-	providerKeeper := s.providerChain.App.(*appProvider.App).ProviderKeeper
-	consumerKeeper := s.consumerChain.App.(*appConsumer.App).ConsumerKeeper
-
-	packet := channeltypes.NewPacket([]byte{}, 1, consumertypes.PortID, s.path.EndpointA.ChannelID,
-		providertypes.PortID, "wrongchannel", clienttypes.Height{}, 0)
-
-	ack := providerKeeper.OnRecvSlashPacket(s.providerCtx(), packet, ccv.SlashPacketData{})
-	s.Require().NotNil(ack)
-
-	err := consumerKeeper.OnAcknowledgementPacket(s.consumerCtx(), packet, channeltypes.NewResultAcknowledgement(ack.Acknowledgement()))
-	s.Require().NoError(err)
-
-	err = consumerKeeper.OnAcknowledgementPacket(s.consumerCtx(), packet, channeltypes.NewErrorAcknowledgement("another error"))
-	s.Require().Error(err)
-}
